requirements.txt
PyYAML==5.3
click==7.1.2